added caching, caffeine and divided logic for role permittions

+++ src/main/java/ru/effectmobile/task_management_system/service/facade/impl/TaskFacadeImpl.java
@@ -9,11 +9,10 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.effectmobile.task_management_system.dto.filters.TaskFilterDTO;
 import ru.effectmobile.task_management_system.dto.requests.TaskRequestDTO;
 import ru.effectmobile.task_management_system.dto.responses.TaskResponseDTO;
-import ru.effectmobile.task_management_system.exception.custom.auth.UserDoesntHaveEnoughRightsException;
 import ru.effectmobile.task_management_system.model.entity.Task;
 import ru.effectmobile.task_management_system.model.entity.User;
-import ru.effectmobile.task_management_system.model.enums.Role;
 import ru.effectmobile.task_management_system.model.metadata.MetaData;
+import ru.effectmobile.task_management_system.service.base.PermissionService;
 import ru.effectmobile.task_management_system.service.base.TaskService;
 import ru.effectmobile.task_management_system.service.base.UserService;
 import ru.effectmobile.task_management_system.service.facade.TaskFacade;
@@ -22,9 +21,6 @@ import ru.effectmobile.task_management_system.service.factory.TaskFactory;
 import ru.effectmobile.task_management_system.service.mapper.TaskMapper;
 
 import java.util.UUID;
-
-import static ru.effectmobile.task_management_system.exception.util.ExceptionMessageUtil.Messages.TASK_CREATE_OR_DELETE_FORBIDDEN_MESSAGE;
-import static ru.effectmobile.task_management_system.exception.util.ExceptionMessageUtil.Messages.TASK_UPDATE_FORBIDDEN_MESSAGE;
 
 @Slf4j
 @Service
@@ -36,6 +32,7 @@ public class TaskFacadeImpl implements TaskFacade {
     private final TaskFactory taskFactory;
     private final UserService userService;
     private final MetaDataFactory metaDataFactory;
+    private final PermissionService permissionService;
 
     @Override
     @Transactional(readOnly = true)
@@ -59,9 +56,11 @@ public class TaskFacadeImpl implements TaskFacade {
     @Override
     @Transactional
     public TaskResponseDTO createTask(TaskRequestDTO taskDTO, String email) {
-        User author = userService.findByEmail(email);
-        canUserCreateOrDeleteTask(email);
+        User user = userService.findByEmail(email);
+        permissionService.checkCanCreateTask(user);
         log.info("Creating a new task with title: {}", taskDTO.title());
+
+        User author = userService.findByEmail(email);
         MetaData metaData = metaDataFactory.createMetaData();
         Task task = taskFactory.createTask(taskDTO, author, metaData);
         setAssignee(task, taskDTO.assigneeId());
@@ -75,8 +74,10 @@ public class TaskFacadeImpl implements TaskFacade {
     @Override
     @Transactional
     public TaskResponseDTO updateTask(UUID id, TaskRequestDTO taskDTO, String email) {
-        canUserUpdateTask(email, taskDTO);
+        User user = userService.findByEmail(email);
+        permissionService.checkCanUpdateTask(user, taskDTO.assigneeId());
         log.info("Updating task with ID: {}", id);
+
         Task task = taskService.findById(id);
         taskMapper.updateTaskFromRequestDTO(taskDTO, task);
         setAssignee(task, taskDTO.assigneeId());
@@ -90,7 +91,8 @@ public class TaskFacadeImpl implements TaskFacade {
     @Override
     @Transactional
     public void deleteTask(UUID id, String email) {
-        canUserCreateOrDeleteTask(email);
+        User user = userService.findByEmail(email);
+        permissionService.checkCanDeleteTask(user);
         log.warn("Deleting task with ID: {}", id);
         taskService.deleteById(id);
         log.info("Task deleted successfully with ID: {}", id);
@@ -116,23 +118,6 @@ public class TaskFacadeImpl implements TaskFacade {
             log.debug("Assigning user ID: {} to task ID: {}", assigneeId, task.getId());
             User assignee = userService.findById(assigneeId);
             task.setAssignee(assignee);
-        }
-    }
-
-    private void canUserUpdateTask(String email, TaskRequestDTO taskDTO) {
-        User user = userService.findByEmail(email);
-        Role role = user.getRole();
-        if (!role.equals(Role.ROLE_ADMIN)) {
-            if (!user.getId().equals(taskDTO.assigneeId())) {
-                throw new UserDoesntHaveEnoughRightsException(TASK_UPDATE_FORBIDDEN_MESSAGE);
-            }
-        }
-    }
-
-    private void canUserCreateOrDeleteTask(String email) {
-        Role role = userService.findByEmail(email).getRole();
-        if (!role.equals(Role.ROLE_ADMIN)) {
-            throw new UserDoesntHaveEnoughRightsException(TASK_CREATE_OR_DELETE_FORBIDDEN_MESSAGE);
         }
     }
 }
